Add getTextGrid utility
For rendering spreadsheet-like grids of text labels. This is a breaking change, as getTextLabel
has been altered to accept the width and height as optional params so that text labels can be
auto-sized based on the text width. Also, createBarGraph has been altered so that entries
can have an optional "icon" property as either a URL or a canvas/image (previously it was just
the "imageUrl" property, only a URL)